Fix for [bug 2094] and [bug 2727].
Rev: src/las.c:1.239
Rev: src/program.c:1.309
Rev: src/program.h:1.115

+++ src/program.h
@@ -5,7 +5,7 @@
 \*/
 
 /*
- * $Id: program.h,v 1.114 2001/02/05 12:04:45 per Exp $
+ * $Id: program.h,v 1.115 2002/02/06 17:26:15 grubba Exp $
  */
 #ifndef PROGRAM_H
 #define PROGRAM_H
@@ -434,7 +434,7 @@ PMOD_EXPORT int add_function_constant(char *name, void (*cfun)(INT32), char * ty
 PMOD_EXPORT int debug_end_class(char *name, ptrdiff_t namelen, INT32 flags);
 INT32 define_function(struct pike_string *name,
 		      struct pike_string *type,
-		      unsigned INT8 flags,
+		      unsigned INT16 flags,
 		      unsigned INT8 function_flags,
 		      union idptr *func,
 		      unsigned INT16 opt_flags);
